get_5_stocks/requirements.txt
finnhub-python
openai